backend/requirements.txt
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.10.0
